% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrandDensity2.R
\name{coord2net2}
\alias{coord2net2}
\title{Add polygon ID to object with long and lat}
\usage{
coord2net2(SpFile2, SpFile2ID, coordFile, lonName, latName)
}
\arguments{
\item{SpFile2}{a spatial polygon dataframe object}

\item{SpFile2ID}{a text field that identifies a varibale that defines polygon in the sp dataframe}

\item{coordFile}{an object with long and lat}

\item{lonName}{name of longitude variable in coordFile}

\item{latName}{name of latitude variable in coordFile}
}
\description{
Overlay spatial polygon dataframe object over object with lon and lat. The function outputs coordFile with SpFile2ID added
}
\examples{
\dontrun{
coord2net2(net_coord_SP_base,"ID",coordFile=someObject,lonName=long,latName=lat)
}
}
