% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrandDensity2.R
\name{cutRange}
\alias{cutRange}
\title{Convert Numeric to Factor based on supplied ranges}
\usage{
cutRange(df, value_col, cut_range, adapt = T)
}
\arguments{
\item{df}{a data frame}

\item{value_col}{a name of numeric column to be converted}

\item{cut_range}{character vector that is simialar to the output of *cut* function. Each element represents a range closed on the right}

\item{adapt}{logical, if TRUE add o zero range to put zero values into it's own range}
}
\description{
cutRanges divides the range of *value_col* and codes the values according to which interval they fall in based on the provided
a new column value_cat to the returned dataframe
}
\examples{
\dontrun{
df2<-cutRange(df,"customer_value",df_ranges,F)
}
}
