% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_coord}
\alias{net_coord}
\title{net_coord is ggplot fortified df
redundant: can be obtained by fortifying net_coord_SP}
\format{dataframe}
\source{
generated data
}
\usage{
net_coord
}
\description{
net_coord is ggplot fortified df
redundant: can be obtained by fortifying net_coord_SP
}
\keyword{datasets}
