% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_coord_centers}
\alias{net_coord_centers}
\title{net_coord_centers: geographical centres of polygon squares.
somewhat redundant data as centers can be recovered from a slot on net_coord_SP_base.
The data covers areas outside agri regions as well.}
\format{dataframe}
\source{
generated data
}
\usage{
net_coord_centers
}
\description{
net_coord_centers: geographical centres of polygon squares.
somewhat redundant data as centers can be recovered from a slot on net_coord_SP_base.
The data covers areas outside agri regions as well.
}
\keyword{datasets}
