% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_coord_SP_base}
\alias{net_coord_SP_base}
\title{net_coord_SP_base: Spatial polygons dataframe.
the data is build around agricultural regions of Canada/Western.
Each polygon in the datasource is approx 11 by 11 km square.}
\format{'sp' object}
\source{
generated data
}
\usage{
net_coord_SP_base
}
\description{
net_coord_SP_base: Spatial polygons dataframe.
the data is build around agricultural regions of Canada/Western.
Each polygon in the datasource is approx 11 by 11 km square.
}
\keyword{datasets}
