% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Net_Postal_base}
\alias{Net_Postal_base}
\title{Net_Postal_base: connect net_coord to Postal Code. The data is for agricultural regions of Canada/Western.}
\format{dataframe}
\source{
generated data
}
\usage{
Net_Postal_base
}
\description{
Net_Postal_base: connect net_coord to Postal Code. The data is for agricultural regions of Canada/Western.
}
\keyword{datasets}
