% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrandDensity.R
\name{valueByNet2Density}
\alias{valueByNet2Density}
\title{Convert value by net_coord to density}
\usage{
valueByNet2Density(df, value_col, resolution = 10, h = 2, n = 50)
}
\arguments{
\item{df}{dataframe with net coord and some value, 2 columns}

\item{value_col}{value, most likely transactions summed by net_coord}

\item{resolution}{how many categories for density will be created. default=10}

\item{h}{comes from MASS::kde2d. kernel parameters. See MASS::kde2d for details}

\item{n}{comes from MASS::kde2d. soze of the grid.See MASS::kde2d for details}
}
\description{
Convert value by net_coord to density
}
